pyproject.toml
[project]
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "DTimperi - CSA", email = "d.timperi@csadocuments.it"},
]
dependencies = ["torch>=2.7.1", "numpy>=2.3.2", "pillow>=11.3.0", "tqdm>=4.67.1", "cryptography>=45.0.5", "accelerate>=1.9.0", "optimum[onnxruntime-gpu]>=1.26.1", "onnxruntime-gpu>=1.22.0", "more-itertools>=10.7.0", "opencv-python>=4.11.0.86"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = true
